Move assets-related tasks into separate file
This code re-defines the assets-related tasks in order to avoid a load
of INFO-level warnings generated by the assets:precompile task [1].

It wasn't immediately obvious to me that was what was happening (even
though I made the change!). I think moving the code into a separate
`lib/tasks/assets.rake` file is more idiomatic and makes it slightly
more obvious what's going on.

[1]: https://github.com/alphagov/signon/pull/2179