Show user status as tag instead of strikethrough
The strikethrough is not part of the design system, so we've decided to
use the govuk-tag style instead.

The design on the card showed green for active users, grey for suspended
users, but there were no examples of locked or invited users.

I've chosen to only use green for active users, and grey for all other
statuses.

At James' suggestion, I created a new method, `status_with_tag`, to
avoid changing the display in other places where `status` is used, such
as the CSV export of the users, and the summary view for user edit pages