Add customised accessible autocomplete
This uses the `accessible-autocomplete` package added in the previous
commits with a few customisations:
- library-internal config as defined in the `init` function
- a listener so that when the value of the autocomplete element no
  longer matches the option selected in the underlying `select` element,
  the value is deselected. This prevents accidental submits of the form
  submitting a value that no longer matches what the user sees
- a clear button. By default, when entering the input with an existing
  value, it won't show you other options, and it's not necessarily
  obvious that you need to delete the current value in order to see the
  list of options again. This adds a clear way of resetting the options
  list (and current value)
- custom styling for the above, as well as (among other things) the
  default blank option so that its height matches the non-blank options

The format of the JavaScript file follows the GDS Way, but some of the
custom logic is taken from
`UKGovernmentBEIS/beis-report-official-development-assistance`

We don't test the main autocomplete functionality here, since it comes
from an external (alphagov) library and is tested there. We do test the
custom behaviour we've added. We found that we needed to wait in a few
places in order for event listener callbacks to fully execute before
asserting on the expected outcomes

Co-authored-by: George Eaton <george@dxw.com>