I18n User#organisation_id blank validation message
I can't find a canonical source of documentation but passing a hash as
the third argument to `ActiveModel::Errors#add`[1] appears to make those
values available for interpolation in the I18n string.

This validation error message is already being tested in "organisation
admin must belong to an organisation" and "super organisation admin must
belong to an organisation" in user_test.rb.

[1]: https://api.rubyonrails.org/classes/ActiveModel/Errors.html#method-i-add