Show Status for API users rather than yes/no
Calum suggested it would be preferable to show API users' "Status"
(instead of "Suspended?") on the API users table, similar to the Users
table.

This allows us to show the status tag as "Active" or "Suspended", in a
similar way to the Users table, to create consistent expectations for
users of the interface.