WIP add Jasmine test for autocomplete
This isn't working but I'm not entirely sure how to work out why given
the complicated govuk-docker setup. I'm unlikely to have time to look
into this properly in the short term

Leaving it as a WIP, but it's possible that we don't need this? I'm not
sure what our approach is to testing custom JavaScript (the autocomplete
is mostly from a library, but with some customisation)