Revert "Improve User validation messages"
This reverts commit 1eb45b8ce50223816ff7ebbe77d997310a1b2576.

The reverted commit left us in an inconsistent state when it came to
displaying User validation error messages. Some of the validation error
messages were standalone (e.g. "Enter an email for the user") while
others only made sense when combined with the attribute (e.g. "Email
can't contain non-ASCII characters").

I attempted to fix the problems with this inconsistency in PR #2717 but
have now run out of time so have agreed with Craig S that it's better to
revert the problematic commit for now.