Replace comment in SessionsController#log_event with test
A test for this (rather obscure) behaviour was added [1] when the comment
was originally added. However, it was later (probably accidentally)
removed in this commit [2].

I've added a new explanatory controller test to demonstrate the relevant
behaviour and deleted the comment, because it's a lot less likely to
stay up-to-date with the code than the test is.

[1]: https://github.com/alphagov/signon/commit/346e611747d56b30103908f658dff06f8b8cf5cc#diff-c848c7502542e7325a2798553871ed708b8b4dcaa3d290e64d53e904bd92687cR36-R42
[2]: https://github.com/alphagov/signon/commit/4098204ce22684b89af9cd68f7113e67c702e5fd