Add migration to clean up `UserApplicationPermission`s
We've been seeing `NoMethodError`s raised when trying to fetch certain
users' current permissions in the permissions update flow

It appears that the `SupportedPermission` with ID `2316` has been
deleted, but that the associated `UserApplicationPermission`s were not,
despite the model callbacks that should destroy dependencies. There are
46 `UserApplicationPermission`s with a `supported_permission_id` of
`2316`. This is currently the only `supported_permission_id` without a
corresponding `SupportedPermission` that appears in any
`UserApplicationPermission`s, so hopefully the missed callbacks were an
isolated incident. This commit therefore adds a migration to deal with
this specific issue rather than making changes to application code. If
the issue resurfaces with another `supported_permission_id`, we can
investigate further

For logging the permissions removal event:
- an initiator is required, so this is set to me. There are other
  `alphagov` `db` files (outside seeds) that find and use a specific
  user
- there's one `UserApplicationPermission` with a `user_id` that doesn't
  belong to an existing `User`, so logging is skipped unless there is a
  user