Add button to delete application supported permissions
It is not currently possible to delete a supported permission through
the user interface.

This has led to code being run on a Rails console when permissions need
to be deleted, which has led to other issues because the wrong method
was invoked (`delete` instead of `destroy`, which didn't delete the
associated `UserApplicationPermission` records).

Therefore adding a delete button (including confirmation view) to allow
to do be done in the user interface.