Make secrets sync zero-configuration.
We used to have a lot of cruft lying around in Signon, but now it's much
tidier. It's now feasible for us to just import the secrets for all
active (non-retired) Doorkeeper applications and all non-suspended API
users.

This removes an awkward step in setting up or renaming an application
which tends to trip people up.