Make autocomplete arrow button work
There's an [ongoing issue][1] with the accessible autocomplete that
stops the arrow from opening the options menu when `showAllValues: true`
is enabled. The `showAllValues: true` option shows all values when
focused on the autocomplete element without having typed anything

I tried getting this to work in Capybara with Selenium in
`granting_permissions_test.rb` and `account_applications_test.rb`, but
the click event listener on the arrow is making it a challenge to
click on it in a test. Per the Jasmine test, you need to dispatch a
click event rather than just using the `click()` method directly. I
tried doing this in the Capybara test with `execute_script` and various
other methods, but couldn't get the right results

Instead, I've opted for a small single function Jasmine test

This also hides the arrow when the autocomplete input is focused.
Initially I wanted to make the arrow stay but have it toggle the menu,
but I was trying to do this by checking whether the input was the active
element, and it's no longer the active element once you click on the
arrow! I didn't want to get into storing state in a variable, as there's
too much under the hood autocomplete functionality that would affect
visibility and therefore make the conditional logic complicated

[1]: https://github.com/alphagov/accessible-autocomplete/issues/202