Allow adding multiple permissions with autocomplete
The new autocomplete form for adding permissions to apps with more than
eight supported permissions redirects users to the applications path
after adding a permission. Sometimes users want to add multiple
permissions at the same time. This offers a middle ground, whereby we
still have a single value autocomplete, but there's an option to be
redirected back to the form after adding each permission

The `if (addButton)` (and the `if (clearButton)` added earlier) allow
the Jasmine tests to pass. The Jasmine tests don't load the full page,
only the autocomplete component, which is not where those buttons are
located