Fix row alignment in applications tables
In 901963c, styles were added to right align the actions in the final
column of the applications tables. This was described as adopting a
Whitehall-style pattern, and it does follow the pattern of wrapping the
applications tables in a `div` with the class
`govuk-table--with-actions` and applying alignment styles to the last
column in the wrapped table, however the actual applied styles are quite
different. In Whitehall, it's a simple `text-align: right`, whereas in
Signon flex-based styling was introduced. The commit message explained
that this was done so that we could have a gap between elements, since
in Signon we have buttons/button links in addition to simple text links

However, since this change, we've had issues with the height of the
cells in the actions column. This has led to some broken/misaligned row
borders under certain conditions (variation in whether cells need to
wrap their text content)

Using flex on the actual table cell appears to be the issue here: we
lose the consistent height that table cells provide for free. In this
commit, we therefore pivot to wrapping the contents of actions cells in
a classed `div` on which the styles are now applied, which retains the
custom alignment and gap for the actions but restores the default table
styling on the cell itself. This fixes the height/border/row alignment
issues

The `flex-direction` is also switched to `column` at smaller viewport
sizes so that we don't end up with very narrow links/buttons