Show view permissions link even when updatable
Previously on the applications pages the view permissions link wouldn't
show if the user could edit permissions. However, we'll soon fix
delegatable permissions, which will mean that publishing managers will
only see delegatable permissions on the update permissions page. It
would be useful for them to be able to see all permissions, even if they
can't grant non-delegatable ones, so we should show the view link
regardless of update authorisation