Don't default to `signin` permission
We need to fix delegatable permissions, and this will involve changes to
the `UserApplicationPermissionPolicy`. Currently everything is based on
the `signin` permission. This undoes that, updates the
`UserApplicationPermission.for` method so that `supported_permission` is
compulsory, and - for now - passes in the `signin` permission where this
is used. This will be iterated on in later commits to gradually fix
delagatable permissions

This required an update to the `UserApplicationPermissionPolicy` tests,
and I restructured them in the process