Fix duplicate `SupportedPermission` class locally
I've occasionally run into a weird error locally along the lines of:

```
SupportedPermission(#35440) expected, got #<SupportedPermission id: 4, application_id: 2, name: "signin", created_at: "2024-06-27 22:24:44.000000000 +0100", updated_at: "2024-06-27 22:24:44.000000000 +0100", delegatable: true, grantable_from_ui: true, default: false> which is an instance of SupportedPermission(#22720) (ActiveRecord::AssociationTypeMismatch)
```

I researched this a little a while back and tried this solution then,
but then worked out a less hacky way to get things working. However,
I've encountered the same issue again recently, so I'm adding this hack
back in in the hope of preventing future errors

I think the issue is that when we create an app in the seeds file,
there's a callback that creates the `signin` supported permission, but
it doesn't properly load the class. Then we call `SupportedPermission`
explicitly to create more supported permissions in the seeds file. This
somehow results in two different versions of the `SupportedPermission`
class, which is only very occasionally an issue