Add and use `Users::ApplicationPolicy`
We want to move towards providing authorisation to edit permissions
based on the current user (granter), the user being managed (grantee),
and whether the given application has _any_ supported permissions.
Currently the `signin` permission is used in place of the latter. This
authorisation will be based on app-level information, not individual
permission-level information. Using a `UserApplicationPermissionPolicy`
to manage such permissions therefore feels a little misleading

We do need to be able to dictate _which_ non-`signin` permissions are
manageable by a given granter, but currently this works in an all or
nothing manner. This isn't/won't be changed by this shift in policy
class. Later updates will need to be made to other policies and/or
non-policy classes to fix this (see the "Access and permissions"
documentation for information on what dictates this)

This commit doesn't change the authorisation for editing permissions,
but the authorisation to grant and revoke app access feel like they
belong in the same place, following the pattern set by
`Account::ApplicationPolicy`. Moving towards a comparable policy setup
for the `Account` and `Users` namespace to reduce complexity and improve
comprehensibility is one of the goals here

Note that the policy has a custom `initialize` method. I did this so
that we could pass in both the user and the application. This was a
benefit of the `UserApplicationPermissionPolicy`, but the improvement in
comprehensibility/similarity across namespaces feels more useful, and
here we won't require an unused association with a specific permission
when building out other policy methods like `edit_permissions?`

I had wanted to remove `UserApplicationPermissionPolicy#create?` here,
but it's still used in the `Users::SigninPermissionsController`. I
wonder if a later change could shift the
`UserApplicationPermissionPolicy` back to being a `signin`-specific
policy, but with that more limited scope made more explicit