Explicitly policy-wall user view permissions link
The applications index page that this is on is not accessible unless the
current user can edit the given user, but this provides an extra check
that can be used when generating the link itself. If anything changes
to the permissions required to view the page, or if this is used
somewhere else, this check will provide some extra security. At the
moment it just defers to the `UserPolicy#edit?` method, which means this
has no impact on what end users can see and do

The relevant applications controller tests are restructured a little
here to make it easier to cover each context