Use ActionMailer deliveries over CapybaraEmail
After migrating to `mail-notify` v2, `Capybara::Node::Email#body`
returned an empty string, we think specifically when using `view_mail`.
We installed a few different `mail-notify` commits and found that this
began in [22cb87b][1]. We managed to work around this by getting the
emails from `ActionMailer::Base.deliveries`

We had wanted to continue using Capybara to get the link from the body,
but in the end we opted for a bit of a regex hack. It turns out that the
email body contains a Markdown link, not a HTML one. CapybaraEmail was
converting the URL from the Markdown link into a HTML link (using a not
too dissimilar regex to the one we've introduced here), which is the
only reason that Capybara could find the link at all!

[1]: https://github.com/dxw/mail-notify/commit/22cb87b3170d84de63fd7ccfe62c5d876505ba69

Co-authored-by: Mike Patrick <mike.patrick1@digital.cabinet-office.gov.uk>