Update `update_permissions_link` helper method
We've already updated the `Account::ApplicationPolicy` and
`Users::ApplicationPolicy`, which are called in
`account_applications_permissions_links` and
`users_applications_permissions_links` respectively. These will
determine whether `update_permissions_link` is called at all. As before,
`users_applications_permissions_links` then has responsibility for
determining whether there are any permissions that the user can manage.
This is updated to use the latest logic

The tests are a little abstracted here. The aim is to keep the three
'happy' path conditions DRY. The method is essentially two parts: a
conditional gate that decides whether to exit early with an empty
string, and a further conditional expression that determines that path
to use. The latter is the same regardless of how you get there. Arguably
this could be a separate method for simpler testing, but it should then
probably be a private one