Combine update and signin retention perm tests
We didn't really have a general 'updated permissions successfully' test
for either controller. The test for the redirect seemed the most like a
'happy path' test in the users permissions controller tests, so this
updates that so that it also asserts on the updated permissions. The
signin retention test then felt a little redundant, so it's removed to
avoid double testing. The account permissions controller tests are
updated to match