Inline user reloads in tests
We're only reading from the user once post-reload in these tests, so we
can shorten the tests a little by inlining the reload

I'm not entirely sure that we need these reloads in any case. In some
cases we were reloading the wrong user before the assertion, but it's
possible that we had false negatives in those tests. Something for
further investigation another time