Redirect from perms edit if no permissions
If there are no permissions that the current user can manage for the
given grantee, we should redirect back to the applications page and
flash a message to that effect

On the applications page, if there are no manageable permissions for a
given app, we don't include the "Update permissions" link. However,
prior to this commit, a user could manually visit the edit permissions
path using their browser's URL bar. On that page, they wouldn't have
seen any permissions to edit, and if they already had the page loaded,
the `update` action would reject any changes to unmanageable permission.
However, this commit makes it so that users will be redirected, meaning
that in this edge case they won't see a confusing mostly empty edit
permissions page