Fix permissions by non-GDS users Rake task
I noticed that the data had a lot of events in which permissions were
added but not listed. Some of the permissions could have been removed,
but there were too many for that to account for all the omissions. It
turns out I was splitting the event logs' messages incorrectly, so the
task was looking for e.g. "Editor" within an array like ["signin", "
Editor"]. This revises the split to account for the
programmatically-added whitespace, which should (correctly) result in
more matches