Add `with_signin_permissions_for` transient param
to API Users built with the Users factory.

We currently do this for non-API users, but this will allow us to
simplify test setup a bit and introduce some more consistency.

We also extract the duplicate behaviour of finding a
Doorkeeper::Application from its name or return the app if it already is
one into a method here.