Move and namespace various integration tests
This tidies up the integration tests folder a little. Some of these are
already informally namespaced by file and class name, just not with the
`::` syntax. Some others are hard to quickly categorise as relating to
one particular area of the app without reading through the test
descriptions, particularly when comparing account and users test specs.
This makes the groupings more explicit at a glance

I did want to formally namespace the applications tests too, but this
results in `uninitialized constant Application (NameError)`, presumably
because the class/namespace isn't pre-established in the app or
elsewhere in the test suite

It looks like the class name of the organisations test should have been
updated when the roles and organisations were split: https://github.com/alphagov/signon/commit/c71035ef6a2440e1a682f3568864187be2467abb

The Users::Applications and Users::GrantingPermissions tests overlap:
the latter is more well maintained, and can probably be renamed and
replace the former, but I'll do this in a separate commit

There are some tests specs that touch both the account and users areas
of the app: I've left these in the parent directory for now, but they
could possibly be broken down later. The API tests could also be grouped
later