Use transient param for factory-built API Users
It looks like I added the `with_signin_permissions_for` transient param
by mistake in fccfa03 (a cherry-picked commit I then removed this
functionality from after closing an earlier PR).

Here we actually use it to allow us to simplify test setup a bit and
introduce some more consistency.

There's a bit of doubt as to whether the `signin` permission should be
required for certain API user actions, so we may want to think about
removing this in future when we understand the functionality slightly
better.