Remove redundant test
This was testing two things:

1. updating permissions for an app other than the one specified by
params[:application_id]
2. updating permissions for an app the granter (current user) doesn't have
access to

The first of these is already covered in `prevent permissions being
added for other apps` below.

The second is irrelevant - all admins and superadmins should be able to
grant permissions to an application as per the ApiUsersPolicy (only
cares that the user is a superadmin), and the SupportedPermissionsPolicy
(allows all permissions to be updated by admins and superadmins),
regardless of whether they have access to the app or not.