Move basic content check tests to controller spec
These were already partially covered by the controller specs, just not
for both the apps you have access to and apps you don't contexts. It's
overkill to cover these basic tests in both the integration and
controller tests. Given they're just testing the rendered content of a
page, and this being something we often test in controller specs in this
repo, the controller specs feel like the most suitable home. This
deduplication will allow us to simplify the integration tests a little

There were no equivalent tests for this in the users integration test
spec

We might review and revise this approach later, but reducing the
duplication/inconsistency for now should make it easier to work with and
simpler to refactor later