Check addition and removal of other app perms
This test only checked that when updating one app, a permission for
another app couldn't be added. This extends it to cover addition and
removal, and includes changes to the app being updated for comparison

This is essentially a copy of the changes (and commit message)
introduced in fbb63855.

We add an `access_token` for application_b here, even though it's not
used, nor does it affect the outcome of the test, as it would be
required for users to update permissions for that app, as the test is
(mistakenly) attempting to do.