Test viewing permissions in controller specs
The integration tests and permissions controller tests were asserting on
slightly different things but with very similar tests. They were both
testing some part of the conditional logic, but it wasn't obvious to me
why each part would belong in either side

The permissions controller tests are updated to cover the following,
which was previously only tested in the integration tests:
- the second column ("Yes"/"No")
- behaviour for both GOV.UK admins and publishing managers

The ordering assertions are now a little more implicit, since they're
well tested in the `SupportedPermissionTest` and `ApplicationTest` specs

The integration tests were also testing that the view permissions link
appeared on the applications index. We previously tested this quite
thoroughly in both the applications controller and
`ApplicationTableHelper` test specs. A recent change left the latter
with responsibility for testing the complex conditional logic, and the
former with just checking they're rendered in a happy path. This
therefore remains well tested elsewhere

Note: we might revisit and refine the integration/controller divide
later, but for now I'm just opting to move towards consistency and
deduplication, which will hopefully make later more opinionated
refactors a little easier