Remove unnecessary test setup
While in practice we do need the app to have permissions in order to
produce an update permissions link, we don't need to create them in the
test setup here

In practice, this method is only called in view templates that are
accessible only to GOV.UK admins and publishing managers. The
`update_permissions_link` method has some conditional logic that checks
that an app has particular types of permissions dependent on which of
those two role groups the user is in, ignoring normal users. We test
this conditional logic in the tests for `update_permissions_link`, so I
don't think it needs retesting in the tests for these methods (doing so
would lead to some heavy context nesting)

In the tests for these methods, the setup is already a little
unrealistic: it uses a normal user, who - as mentioned above - wouldn't
normally access view templates in which these methods are called.
However, the conditional logic we want to test in this method relies
solely on the output of policy methods, not users' roles, so I think
this is okay. In using a normal user, we effectively bypass the
role-based conditional logic in `update_permissions_link`, which means
we also don't need any permissions on the app (sidenote: a
non-delegatable permission would only work for GOV.UK admins, not
publishing managers, due to the `update_permissions_link` conditional
logic)

So ultimately these tests assume that the app has permissions that the
current user can manage, which I think is reasonable given the
`update_permissions_link` tests cover the other cases