Avoid retesting helper logic in controller tests
We've got some heavy nesting going on in our applications controller
tests at the moment, which makes them a bit unwieldy. I noticed that
we're effectively duplicating nested contexts from the
`ApplicationTableHelper` tests, which are in effect testing conditional
logic from that helper's methods. I think we can assume that the unit
tests for that helper will cover its own conditional logic, and we
should then just test one happy helper path in the controller tests

Note that the we do set up some of the policy requirements to reach the
happy helper paths, but we don't give the app any permissions, which in
practice would be required to reach these paths. However, like in some
of the helper tests for its public methods, we use a normal user, which
effectively bypasses some of the has manageable permissions checks in
one of the helper's private methods. We're setting things up a little
unrealistically in other areas, e.g. by opting for policy stubbing with
a normal user rather than using a user with a more privileged role, so I
think this is okay for now, but it's worth noting in case something
changes that requires us to beef up the test setup in future