Create new structure for access/perms int. tests
This is a new consistent structure for the access and permissions tests.
The actual test implementations will be ported over from the existing
access and permissions tests gradually in later commits

The specs are broken into three given the number of contexts that are
covered. I'm not entirely convinced about how comprehensive to go with
the condition testing, but this feels like a decent level of
documentation and aligns reasonably well with the level in the
associated Markdown documents (docs/access_and_permissions.md). This
condition testing is well covered by unit tests, but it feels useful to
cover it here too for those reasons

Small note: I've ungrammatically gone for "a organisation_admin" in a
few test descriptions to keep things consistent with descriptions
generated in loops

I'd like to keep things DRY within specs using shared methods for
"should be able to" and "should not be able to" tests (i.e. two methods)
and potentially between namespaces if there's any setup or set of steps
that are the same regardless of whether your managing your own or
others' access and permissions (shared methods can go in a support file)

Given the tests for updating permissions for apps with more than eight
permissions are essentially identical other than minor differences in
setup, I've consolidated those into a single spec