Group tests in contexts
Here we group the happy path tests in one context, and the
less-than-happy tests in another.

I wasn't 100% sure on where to draw the line with "when updating
permissions for app A, prevent additionally adding or removing
permissions for app B" as it's a mix of happy and unhappy paths, but I'm
hoping it fits in the way I've worded the context!