Move integration tests for 9+ permissions apps
These are the tests for granting permissions for apps with nine or more
permissions. Instead of making the account and users updating
permissions specs really long and duplicative, I opted for having a
single spec with a grantee who could either be the current user or
another user. The tests here are even more alike between namespaces than
tests for apps with eight or fewer permissions, which had some variation
based on permissions/context

Because the JavaScript tests need to set the JavaScript driver before
running the setup (in particular logging in), the setup shared between
JavaScript and non-JavaScript tests is encapsulated in a method rather
than a `setup` block