Upgrade jasmine-browser-runner dependencies
to fix a security vulnerability.

We had a medium security vulnerability because jasmine-browser-runner
depends on express 4.20.0 which depends on serve-static 0.16.0 which
depends on send 0.18.0, which is currently vulnerable due to
https://github.com/advisories/GHSA-m6fv-jmcg-4jfg. We pinned the express
dependency to resolve this but now that a new version of serve-static
has been released, we can unpin this.

The changes to `yarn.lock` were generated by running `yarn upgrade
jasmine-browser-runner` to ensure it's using the patched version
(0.19.0) of the `send` dependency.