Make Account::RemovingAccessTests more verbose
I believe these tests are a little too DRY at the expense of
readability. Although DRY is a useful standard to strive for in
implementation code, in Integration tests it's less valuable as it's
useful to have the individual steps written out and easy to read at a
glance to serve as documentation for how the software should work.

Although the code for the three user journeys (Account, Users and API
Users) is very similar and can possibly be extracted into some helper
classes, at the moment the journeys are slightly different, so having
test helpers across a few separate files with conditional flags adds a
fair amount of overhead on the reader when understanding where to add
new tests.

Because this shared test set up is re-used across multiple journeys, it
also makes it harder to introduce new functionality in smaller chunks.
When writing a tweak to logic for one user journey (e.g.  Accounts) I
also had to update the logic for the other (Users) to avoid having
failing tests on a commit, resulting in more changes being introduced at
once.

Although it's a a menial task to have to update this logic in multiple
places when making changes, I believe the benefits in keeping the
tests simple and readable outweigh the pain of writing similar tests
across multiple files.