Add success banner when users are granted access to an app
This was a lot more complex than I'd hoped due to the difficulties in
cleanly distinguishing between a user having the Signin permission
granted vs having all permissions other than the Signin permission
removed.

I hope this is reasonably clear and duplication-free. I'd rather have
avoided passing the `granting_access` flag if possible and opting for
something slightly more extendable in future, but it won out over the
following alternative approaches:

1. Performing a lot of this logic in the view - harder to test, a bit
   too much duplication and adding too much logic to a rendering
   template.
2. Refactoring the `ApplicationPermissionsHelper` to support a new path
   for granting access - this had a lot of fairly complex branching
   logic in it and felt fairly geared towards "permissions" rather than
   "access" so I've opted against extending it further.

I've opted against writing controller tests for this behaviour. Although
we have some where we test for the presence of `flash[:application_id]`,
this feels a bit over the top, and I think it's sufficiently covered
with integration tests that are testing the view.